Statically include android.test.legacy
Replaces a runtime dependency on android.test.runner and
android.test.base libraries with a static dependency on
android.test.legacy.

This is mainly to provide compile time access to the hidden
LaunchPerformanceBase class without adding a dependency on internal
APIs. It is also a tiny step on the way to removing all dependencies on
those runtime libraries which will allow them to be removed completely.

Bug: 69899800
Bug: 30188076
Test: m -j checkbuild
Change-Id: Ied37beaafe47820fdbf44e957e671c3e7258ddd3